refactor: 실수형 타입 변경

+++ src/main/java/blackjack/domain/player/BetMoney.java
@@ -5,7 +5,7 @@ import blackjack.exception.NotEnoughBetAmountException;
 public class BetMoney {
 
     private static final long MIN_BETTING_MONEY = 1;
-    private static final float BLACKJACK_BONUS = 1.5f;
+    private static final double BLACKJACK_BONUS = 1.5;
 
     private final long betMoney;
 
@@ -24,7 +24,7 @@ public class BetMoney {
         return betMoney;
     }
 
-    public float getBlackjackBetMoney() {
+    public double getBlackjackBetMoney() {
         return betMoney * BLACKJACK_BONUS;
     }
 }
